added search to tab bar controller and added tab bar icons

+++ yesand/RateView.m
@@ -130,8 +130,8 @@
 - (void)handleTouchAtLocation:(CGPoint)touchLocation {
     if (!self.editable) return;
 
-    unsigned long newRating = 0;
-    for(unsigned long i = self.imageViews.count - 1; i >= 0; i--) {
+    int newRating = 0;
+    for(int i = self.imageViews.count - 1; i >= 0; i--) {
         UIImageView *imageView = [self.imageViews objectAtIndex:i];
         if (touchLocation.x > imageView.frame.origin.x) {
             newRating = i + 1;
